+++ test/client.spec.ts
@@ -2,38 +2,153 @@ import fetchMock from 'fetch-mock'
 import { expect } from 'chai'
 import { client, oauth, sdkVersionHeader } from '../dist/index'
 
+const company = { id: 4 };
 const project = { id: 3 };
 const me = { id: 42, login: 'foo@procore.com', name: 'foo' };
 const rfi = { id: 1, subject: 'Create RFI Subject', assignee_id: 2945 };
 const idsToDelete = [{ id: 1 }, { id: 2 }];
 const token = 'token';
 const hostname = 'https://app.procore.com';
-const headers = { 
-    'Accept': 'application/json', 
-    'Content-Type': 'application/json', 
-    'Procore-Sdk-Version': sdkVersionHeader, 
-    'Procore-Sdk-Language': 'javascript' 
+const headers = {
+  Accept: 'application/json',
+  'Content-Type': 'application/json',
+  'Procore-Sdk-Version': sdkVersionHeader,
+  'Procore-Sdk-Language': 'javascript',
 };
 
 describe('client', () => {
-  it('uses a custom formatter', async () => {
-    const authorizer = oauth(token);
-    const procore = client(authorizer);
-    let counter = 1;
+  beforeEach(() => {
+    fetchMock.reset();
+  });
 
-    function formatter(response: any) {
-      counter += 1;
-      return response.json();
-    }
+  context('Override fetch', () => {
+    it('uses a custom formatter', async () => {
+      const authorizer = oauth(token);
+      const procore = client(authorizer);
+      let counter = 1;
 
-    fetchMock.get(`${hostname}/foo/me`, me);
+      function formatter(response: any) {
+        counter += 1;
+        return response.json();
+      }
 
-    const { body } = await procore.get('/foo/me', { formatter });
+      fetchMock.get(`${hostname}/foo/me`, me);
 
-    expect(body).to.eql(me);
-    expect(counter).to.eql(2);
-    fetchMock.restore();
-  })
+      const { body } = await procore.get('/foo/me', { formatter });
+
+      expect(body).to.eql(me);
+      expect(counter).to.eql(2);
+      fetchMock.restore();
+    });
+
+    it('sets company id header using clientOptions', async () => {
+      const authorizer = oauth(token);
+      const procore = client(authorizer, { headers: { ...headers } }, { defaultCompanyId: company.id });
+
+      const customHeaders = {
+        'Procore-Company-Id': `${company.id}`,
+      };
+
+      fetchMock.get(
+        {
+          url: `${hostname}/foo/projects`,
+          headers: {
+            ...headers,
+            ...customHeaders,
+          },
+        },
+        project
+      );
+
+      const { body } = await procore.get('/foo/projects');
+
+      expect(body).to.eql(project);
+      fetchMock.restore();
+    });
+
+    it('sets company id header using requestConfig', async () => {
+      const authorizer = oauth(token);
+      const procore = client(authorizer, { headers: { ...headers } });
+
+      const customHeaders = {
+        'Procore-Company-Id': `${company.id}`,
+      };
+
+      fetchMock.get(
+        {
+          url: `${hostname}/foo/projects`,
+          headers: {
+            ...headers,
+            ...customHeaders,
+          },
+        },
+        project
+      );
+
+      const { body } = await procore.get('/foo/projects', {
+        companyId: company.id
+      });
+
+      expect(body).to.eql(project);
+      fetchMock.restore();
+    });
+
+    it('sets company id header using requestConfig with different clientOptions', async () => {
+      const authorizer = oauth(token);
+      const procore = client(authorizer, { headers: { ...headers } }, { defaultCompanyId: 6765 });
+
+      const customHeaders = {
+        'Procore-Company-Id': `${company.id}`,
+      };
+
+      fetchMock.get(
+        {
+          url: `${hostname}/foo/projects`,
+          headers: {
+            ...headers,
+            ...customHeaders,
+          },
+        },
+        project
+      );
+
+      const { body } = await procore.get('/foo/projects', {
+        companyId: company.id
+      });
+
+      expect(body).to.eql(project);
+      fetchMock.restore();
+    });
+
+    it('uses a custom header', async () => {
+      const authorizer = oauth(token);
+      const procore = client(authorizer, { headers: { ...headers } });
+
+      const customHeaders = {
+        'Procore-Company-Id': `${company.id}`,
+      };
+
+      fetchMock.get(
+        {
+          url: `${hostname}/foo/projects`,
+          headers: {
+            ...headers,
+            ...customHeaders,
+          },
+        },
+        project
+      );
+
+      const { body } = await procore.get('/foo/projects', {
+        headers: {
+          ...customHeaders,
+        },
+      });
+
+      expect(body).to.eql(project);
+      fetchMock.restore();
+    });
+  });
 
   context('using oauth', () => {
     describe('request defaults', () => {
@@ -48,7 +163,7 @@ describe('client', () => {
         fetchMock.restore();
       })
 
-      it('allows headers override', async () => {
+      it('allows response headers override', async () => {
         const headers = new Headers();
         const authorizer = oauth(token);
         const procore = client(authorizer, { headers });
@@ -180,7 +295,7 @@ describe('client', () => {
 
       it('creates a resource', async () => {
         fetchMock.post(
-            { url: `${hostname}/rest/v1.0/projects/${project.id}/rfis`, headers: headers }, rfi);
+          { url: `${hostname}/rest/v1.0/projects/${project.id}/rfis`, headers: headers }, rfi);
 
         const { body } = await procore
           .post({
@@ -196,7 +311,7 @@ describe('client', () => {
         fetchMock.post(
           { url: `${hostname}/rest/v1.0/projects/${project.id}/rfis`, headers: headers }, (url, opts: RequestInit) => {
             return opts.body;
-        });
+          });
 
         const { body } = await procore
           .post({
@@ -318,7 +433,7 @@ describe('client', () => {
         fetchMock.delete(
           { url: `${hostname}/rest/v1.0/projects/${project.id}/rfis/${rfi.id}`, headers: headers }, (url, opts: RequestInit) => {
             return { body: opts.body, status: 200 };
-        });
+          });
 
         const { body } = await procore
           .delete({