+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 3.0.2 (September 2022)
+
+* Add support for adding `Procore-Company-Id` request header when `defaultCompanyId` is passed in `ClientOptions`.
+* Add support for adding `Procore-Company-Id` request header when `companyId` is passed in `RequestConfig`.
+* Add support for customer headers.
+
 ## 3.0.1 (June 2021)
 
 * Adds support for Rest API versioning (rest/v1.0 is)